requirements.txt
Django>=4.2,<6.0
django-environ>=0.11
psycopg[binary]>=3.1
